Modify web runner state management
Only one request at a time has read/write access to a particular
session and any additional request requiring any session access is
blocked until the previous request is completed.

So now the runner uses memory cache to store the state and the handler
has read only access to the session to be able to read the session id
and use it as a key in the cache object.