Complete ConsolidatedRunner that will process multiple assemblies, both nunit and xunit.
Xunit will no longer output text as a whole - just for the individual tests.  It appears outputting it all as one was a workaround for nunit.  Or perhaps we can switch nunit to attach text output to each individual test too.

Rename assemblies and namespaces to reflect more generic nature of the test runner now (not nunit specific).